chore: Moved app provider to `@/components`
